# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Risk Radar",
    "description": "Real-time transaction risk intelligence. **Advisory**: this service produces a decision (ALLOW / MONITOR / REVIEW / HOLD) with its reasons. It does not block, hold, reverse or freeze anything \u2014 enforcement is the caller's responsibility and is documented as such in the contract.",
    "version": "1.1.0"
  },
  "paths": {
...
    "/v1/cases/{case_id}/workflow": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Case Workflow",
        "description": "Stage, lifecycle and steps for one case.",
        "operationId": "case_workflow_v1_cases__case_id__workflow_get",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Case Workflow V1 Cases  Case Id  Workflow Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/actions": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Record Action",
        "description": "A step done in the bank's systems, recorded here with its result.\n\nOnly the person working the case records steps on it, so the record says who\ndid what. Risk Radar does not perform the step (D7); it remembers it.",
        "operationId": "record_action_v1_cases__case_id__actions_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaseActionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Record Action V1 Cases  Case Id  Actions Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/cases/{case_id}/return": {
      "post": {
        "tags": [
          "workflow"
        ],
        "summary": "Return Case",
        "description": "Hand an escalated case back to the analyst who raised it, with findings.",
        "operationId": "return_case_v1_cases__case_id__return_post",
        "parameters": [
          {
            "name": "case_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Case Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReturnIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Return Case V1 Cases  Case Id  Return Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflow/pipeline": {
      "get": {
        "tags": [
          "workflow"
        ],
        "summary": "Pipeline",
        "description": "Every case in its stage, oldest first, with how long it has waited there.\n\nD94: \"every case\" means every case this person may see.",
        "operationId": "pipeline_v1_workflow_pipeline_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 40,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Pipeline V1 Workflow Pipeline Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CaseActionIn": {
        "properties": {
          "action_code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 3,
            "title": "Action Code"
          },
          "result": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Result"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "action_code",
          "result"
        ],
        "title": "CaseActionIn",
        "description": "D83: a recommended step done in the bank's systems, recorded with its result."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReturnIn": {
        "properties": {
          "findings": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 5,
            "title": "Findings"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "findings"
        ],
        "title": "ReturnIn",
        "description": "D83: an escalated case handed back to the analyst who raised it."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}